requirements.txt
Pillow
torch==2.4.0
transformers
pdf2image
pyppeteer
